+++ config/initializers/event_dispatcher.rb
@@ -1,1 +1,0 @@
-EventDispatcher.disable_send_message! if Rails.env.in? %w(development test playground)